requirements.txt
arabic-reshaper==3.0.0
certifi==2024.6.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.0
guidata==3.1.0
guiqwt==4.4.4
h5py==3.11.0
idna==3.7
kiwisolver==1.4.5
matplotlib==3.9.0
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
pyparsing==3.1.2
pyqt-tools==1.0.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-bidi==0.4.2
python-dateutil==2.9.0.post0
PythonQwt==0.12.5
pytz==2024.1
QtPy==2.4.1
requests==2.32.3
scipy==1.14.0
six==1.16.0
tomli==2.0.1
tzdata==2024.1
urllib3==2.2.2
